feature: common plot functions were separated into a new file to improve maintainability
