+++ plot_observed_category.py
@@ -10,19 +10,9 @@ import cartopy.crs as ccrs
 import cartopy.feature
 from pathlib import Path
 import configuration
+from plot_common_functions import manipular_nc
 
 cfg = configuration.Config.Instance()
-
-def manipular_nc(archivo, variable, lat_name, lon_name, lats, latn, lonw, lone):
-    """gets netdf variables"""
-    #reportar lectura de un archivo descargado
-    cfg.report_input_file_used(archivo)
-    #continuar ejecución
-    dataset = xr.open_dataset(archivo, decode_times=False)
-    var_out = dataset[variable].sel(**{lat_name: slice(lats, latn), lon_name: slice(lonw, lone)})
-    lon = dataset[lon_name].sel(**{lon_name: slice(lonw, lone)})
-    lat = dataset[lat_name].sel(**{lat_name: slice(lats, latn)})
-    return var_out, lat, lon
 
 def main(args):
     """Plot observed category"""
